+++ projects/rccl/test/transport/NetIbMPI/NetIbMPITestBase.hpp
@@ -1,0 +1,463 @@
+/*************************************************************************
+ * Copyright (c) 2025 Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * See LICENSE.txt for license information
+ ************************************************************************/
+
+#pragma once
+
+#include <gtest/gtest.h>
+#include <hip/hip_runtime.h>
+#include "MPITestBase.hpp"
+#include "ResourceGuards.hpp"
+#include "TestChecks.hpp"
+#include "DeviceBufferHelpers.hpp"
+#include "HostBufferHelpers.hpp"
+#include "nccl.h"
+#include "net.h"
+#include <vector>
+#include <memory>
+#include <cstring>
+#include <cstdarg>
+#include <cstdio>
+#include <cstdlib>
+#include <algorithm>
+
+#ifdef MPI_TESTS_ENABLED
+
+// Import helper namespaces
+using namespace RCCLTestGuards;
+using namespace RCCLTestHelpers;
+
+// External NET IB plugin
+extern ncclNet_t ncclNetIb;
+// External NET IB-CAST plugin (WRR scheduler, multi-QP)
+extern ncclNet_t netIbCast;
+
+// NET IB-specific resource deleters
+struct NetMHandleDeleter {
+    ncclNet_t* net;
+    void* comm;
+
+    NetMHandleDeleter(ncclNet_t* n = nullptr, void* c = nullptr) : net(n), comm(c) {}
+
+    void operator()(void* mhandle) const {
+        if (mhandle && net && comm) {
+            int rank = MPIEnvironment::world_rank;
+            TEST_INFO("Rank %d: NetMHandleDeleter - Deregistering memory handle (mhandle=%p, comm=%p)",
+                      rank, mhandle, comm);
+            ncclResult_t result = net->deregMr(comm, mhandle);
+            TEST_INFO("Rank %d: NetMHandleDeleter - deregMr result: %d", rank, result);
+        }
+    }
+};
+
+// NET IB connection guard
+class NetConnectionGuard {
+private:
+    ncclNet_t* net_;
+    void* sendComm_;
+    void* recvComm_;
+    void* listenComm_;
+
+public:
+    explicit NetConnectionGuard(ncclNet_t* net)
+        : net_(net), sendComm_(nullptr), recvComm_(nullptr), listenComm_(nullptr) {}
+
+    ~NetConnectionGuard() {
+        if (sendComm_ && net_) {
+            net_->closeSend(sendComm_);
+        }
+        if (recvComm_ && net_) {
+            net_->closeRecv(recvComm_);
+        }
+        if (listenComm_ && net_) {
+            net_->closeListen(listenComm_);
+        }
+    }
+
+    void setSendComm(void* comm) { sendComm_ = comm; }
+    void setRecvComm(void* comm) { recvComm_ = comm; }
+    void setListenComm(void* comm) { listenComm_ = comm; }
+
+    NetConnectionGuard(const NetConnectionGuard&) = delete;
+    NetConnectionGuard& operator=(const NetConnectionGuard&) = delete;
+};
+
+// Type alias for NetMHandleGuard using ResourceGuard
+using NetMHandleGuard = RCCLTestGuards::ResourceGuard<void*, NetMHandleDeleter>;
+
+// Test fixture for NET IB tests
+class NetIbMPITest : public MPITestBase {
+protected:
+    static constexpr int kMinProcessesForMPI = 2;
+    static constexpr bool kRequirePowerOfTwo = true;
+    static constexpr int kNoNodeLimit = MPITestConstants::kNoNodeLimit;
+
+    // Buffer pattern constants
+    static constexpr int kBytePatternModulo = 256;
+
+    // Timing constants
+    static constexpr int kDefaultTimeoutMs = 5000;
+    static constexpr int kLargeTransferTimeoutMs = 30000;
+    static constexpr int kPollIntervalUs = 10000;  // 10ms
+    static constexpr int kPollIntervalMs = 10;
+    static constexpr int kMaxRetryAttempts = 1000;  // For NULL request handling
+
+    // Buffer size constants
+    static constexpr size_t kSmallBufferSize = 4096;
+    static constexpr size_t kLargeBufferSize = 16 * 1024 * 1024;  // 16 MB
+
+    // Test seed constants
+    static constexpr int kBaseSeedOffset = 1000;
+    static constexpr int kMultiSizeSeedOffset = 2000;
+
+    // Debug output constants
+    static constexpr int kNumDebugSamples = 4;
+
+    // Invalid device ID offset for negative tests
+    static constexpr int kInvalidDeviceOffset = 100;
+
+    // Process count constants
+    static constexpr int kExactTwoProcesses = 2;
+    static constexpr int kMinGpusPerNode = 1;
+
+    // Transfer test constants
+    static constexpr int kNumSequentialTransfers = 100;
+    static constexpr int kTransferTagBase = 300;
+
+    // Timeout constants
+    static constexpr int kLargeTransferTimeout = 30000;
+
+    ncclNet_t* net_;
+    int numDevices_;
+    std::vector<int> deviceIds_;
+    void* initCtx_;
+
+    void SetUp() override {
+        MPITestBase::SetUp();
+        net_ = &ncclNetIb;
+        numDevices_ = 0;
+        initCtx_ = nullptr;
+    }
+
+    void TearDown() override {
+        if (initCtx_) {
+            net_->finalize(initCtx_);
+            initCtx_ = nullptr;
+        }
+        MPITestBase::TearDown();
+    }
+
+    // Helper: Initialize NET IB plugin
+    ncclResult_t InitNetIb() {
+        ncclNetCommConfig_t commConfig = {};
+        commConfig.trafficClass = NCCL_NET_TRAFFIC_CLASS_UNDEF;
+        return net_->init(&initCtx_, 0, &commConfig, nullptr, nullptr);
+    }
+
+    // Helper: Get number of devices
+    ncclResult_t GetDeviceCount(int* ndev) {
+        return net_->devices(ndev);
+    }
+
+    // Helper: Get device properties
+    ncclResult_t GetDeviceProperties(int dev, ncclNetProperties_t* props) {
+        return net_->getProperties(dev, props);
+    }
+
+    // Helper: Create listen comm
+    ncclResult_t CreateListenComm(int dev, ncclNetHandle_t* handle, void** listenComm) {
+        return net_->listen(initCtx_, dev, handle, listenComm);
+    }
+
+    // Helper: Connect to remote
+    ncclResult_t ConnectToRemote(int dev, ncclNetHandle_t* handle, void** sendComm) {
+        return net_->connect(initCtx_, dev, handle, sendComm, nullptr);
+    }
+
+    // Helper: Accept connection
+    ncclResult_t AcceptConnection(void* listenComm, void** recvComm) {
+        return net_->accept(listenComm, recvComm, nullptr);
+    }
+
+    // Helper: Register memory
+    ncclResult_t RegisterMemory(void* comm, void* data, size_t size, int type, void** mhandle) {
+        return net_->regMr(comm, data, size, type, mhandle);
+    }
+
+    // Helper: Register DMA-BUF memory
+    ncclResult_t RegisterDmaBufMemory(void* comm, void* data, size_t size, int type,
+                                      uint64_t offset, int fd, void** mhandle) {
+        return net_->regMrDmaBuf(comm, data, size, type, offset, fd, mhandle);
+    }
+
+    // Helper: Deregister memory
+    ncclResult_t DeregisterMemory(void* comm, void* mhandle) {
+        return net_->deregMr(comm, mhandle);
+    }
+
+    // Helper: Post send operation
+    ncclResult_t PostSend(void* sendComm, void* data, size_t size, int tag,
+                         void* mhandle, void** request) {
+        return net_->isend(sendComm, data, size, tag, mhandle, nullptr, request);
+    }
+
+    // Helper: Post recv operation
+    ncclResult_t PostRecv(void* recvComm, int n, void** data, size_t* sizes,
+                         int* tags, void** mhandles, void** request) {
+        return net_->irecv(recvComm, n, data, sizes, tags, mhandles, nullptr, request);
+    }
+
+    // Helper: Flush operation
+    ncclResult_t FlushRecv(void* recvComm, int n, void** data, int* sizes,
+                          void** mhandles, void** request) {
+        return net_->iflush(recvComm, n, data, sizes, mhandles, request);
+    }
+
+    // Helper: Test request completion
+    // No implementation for this method in the NET IB plugin
+    ncclResult_t TestRequest(void* request, int* done, int* sizes) {
+        return net_->test(request, done, sizes);
+    }
+
+    // Helper: Close send comm
+    ncclResult_t CloseSendComm(void* sendComm) {
+        return net_->closeSend(sendComm);
+    }
+
+    // Helper: Close recv comm
+    ncclResult_t CloseRecvComm(void* recvComm) {
+        return net_->closeRecv(recvComm);
+    }
+
+    // Helper: Close listen comm
+    ncclResult_t CloseListenComm(void* listenComm) {
+        return net_->closeListen(listenComm);
+    }
+
+    // Helper: Make virtual device
+    ncclResult_t MakeVirtualDevice(int* dev, ncclNetVDeviceProps_t* props) {
+        return net_->makeVDevice(dev, props);
+    }
+
+    // Helper: Setup connection between two ranks
+    struct ConnectionPair {
+        void* sendComm = nullptr;
+        void* recvComm = nullptr;
+        void* listenComm = nullptr;
+        ncclNetHandle_t handle;
+    };
+
+    ncclResult_t SetupConnection(int dev, ConnectionPair& pair, int rank, int peerRank) {
+        if (rank == 0) {
+            // Rank 0: Listen
+            RCCL_TEST_CHECK(CreateListenComm(dev, &pair.handle, &pair.listenComm));
+
+            // Send handle to peer
+            MPI_Send(&pair.handle, sizeof(ncclNetHandle_t), MPI_BYTE, peerRank, 0, MPI_COMM_WORLD);
+
+            // Accept connection
+            int done = 0;
+            while (!done) {
+                ncclResult_t result = AcceptConnection(pair.listenComm, &pair.recvComm);
+                if (result == ncclSuccess && pair.recvComm != nullptr) {
+                    done = 1;
+                }
+            }
+        } else {
+            // Rank 1: Connect
+            MPI_Recv(&pair.handle, sizeof(ncclNetHandle_t), MPI_BYTE, peerRank, 0, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
+
+            // Connect to peer
+            int done = 0;
+            while (!done) {
+                ncclResult_t result = ConnectToRemote(dev, &pair.handle, &pair.sendComm);
+                if (result == ncclSuccess && pair.sendComm != nullptr) {
+                    done = 1;
+                }
+            }
+        }
+
+        MPI_Barrier(MPI_COMM_WORLD);
+        return ncclSuccess;
+    }
+
+    // Helper: Initialize device buffer with pattern using DeviceBufferHelpers
+    hipError_t InitializeBuffer(void* buffer, size_t size, int pattern) {
+        // Use template-based helper with custom pattern: (pattern + i) % kBytePatternModulo
+        return initializeBufferWithPattern<uint8_t>(
+            buffer, size,
+            [pattern](size_t i) { return static_cast<uint8_t>((pattern + i) % kBytePatternModulo); }
+        );
+    }
+
+    // Helper: Verify device buffer pattern using DeviceBufferHelpers
+    bool VerifyBuffer(void* buffer, size_t size, int pattern) {
+        // Use template-based helper with pattern verification
+        return verifyBufferData<uint8_t>(
+            buffer, size,
+            [pattern](size_t i) {
+                return static_cast<uint8_t>((pattern + i) % kBytePatternModulo);
+            }
+        );
+    }
+
+    // Helper: Fill host buffer with pattern using HostBufferHelpers
+    void FillHostBuffer(void* buffer, size_t size, int seed) {
+        fillHostBufferWithPattern<uint8_t>(
+            buffer, size,
+            [seed](size_t i) { return static_cast<uint8_t>((seed + i) % kBytePatternModulo); }
+        );
+    }
+
+    // Helper: Verify host buffer pattern using HostBufferHelpers
+    bool VerifyHostBuffer(const void* buffer, size_t size, int seed) {
+        return verifyHostBufferData<uint8_t>(
+            buffer, size,
+            [seed](size_t i) { return static_cast<uint8_t>((seed + i) % kBytePatternModulo); }
+        );
+    }
+
+    // Helper: Retry until the receiver's FIFO slot is ready.
+    void PostSendWithRetry(void* sendComm, void* data, size_t size, int tag,
+                           void* mhandle, void** request) {
+        int attempts = 0;
+        do {
+            ncclResult_t result = PostSend(sendComm, data, size, tag, mhandle, request);
+            ASSERT_EQ(result, ncclSuccess);
+            if (*request != nullptr) break;
+            if (++attempts >= kMaxRetryAttempts) {
+                FAIL() << "PostSend returned NULL request after " << kMaxRetryAttempts << " attempts";
+            }
+            usleep(kPollIntervalUs);
+        } while (*request == nullptr);
+    }
+
+    // Helper: Wait for request completion with timeout
+    ncclResult_t WaitForCompletion(void* request, int* sizes, int timeoutMs = kDefaultTimeoutMs) {
+        if (!request) return ncclInternalError;
+
+        int done = 0;
+        int attempts = 0;
+        const int maxAttempts = timeoutMs / kPollIntervalMs;
+
+        while (!done && attempts < maxAttempts) {
+            ncclResult_t result = TestRequest(request, &done, sizes);
+
+            if (result != ncclSuccess) {
+                return result;
+            }
+
+            if (done) {
+                break;
+            } else {
+                usleep(kPollIntervalUs); // 10ms
+                attempts++;
+            }
+        }
+
+        return done ? ncclSuccess : ncclInternalError;
+    }
+
+    // Composite block: Init plugin + assert device count > 0.
+    // Pass a non-null pointer to receive the count; pass nullptr to discard it.
+    void AssertInitAndGetDevices(int* ndev) {
+        int local = 0;
+        int* p = ndev ? ndev : &local;
+        ASSERT_EQ(InitNetIb(), ncclSuccess);
+        ASSERT_EQ(GetDeviceCount(p), ncclSuccess);
+        ASSERT_GT(*p, 0);
+    }
+
+    // Composite block: SetupConnection + wire up NetConnectionGuard for RAII cleanup.
+    // dev: device index. Uses world_rank to determine listener vs connector.
+    void SetupConnectionWithGuard(int dev, ConnectionPair& pair,
+                                  NetConnectionGuard& guard) {
+        const int rank     = MPIEnvironment::world_rank;
+        const int peerRank = (rank + 1) % 2;
+        ASSERT_EQ(SetupConnection(dev, pair, rank, peerRank), ncclSuccess);
+        if (rank == 0) {
+            guard.setRecvComm(pair.recvComm);
+            guard.setListenComm(pair.listenComm);
+        } else {
+            guard.setSendComm(pair.sendComm);
+        }
+    }
+
+    // Composite block: Post a single irecv. Wraps the 4-array boilerplate.
+    void PostSingleRecv(void* recvComm, void* buf, size_t size, int tag,
+                        void* mhandle, void** request) {
+        void*  bufs[1]    = {buf};
+        size_t sizes[1]   = {size};
+        int    tags[1]    = {tag};
+        void*  handles[1] = {mhandle};
+        ASSERT_EQ(PostRecv(recvComm, 1, bufs, sizes, tags, handles, request), ncclSuccess);
+    }
+
+    // Helper: create a merged device from N physical NICs.
+    // Returns merged device index, or -1 if not enough devices / merge failed.
+    // physDevs: indices of physical (non-merged) devices
+    // props: device properties (indexed by device index)
+    // nNicsToMerge: how many NICs to merge (e.g., 2, 3, 4)
+    // rank: MPI rank of this process
+    int CreateMergedDevice(int nNicsToMerge, int rank, int offset = 0)
+    {
+        if (nNicsToMerge <= 0 || nNicsToMerge > NCCL_NET_MAX_DEVS_PER_NIC) {
+            fprintf(stderr,
+                    "Rank %d requested invalid merge size %d (valid range: 1..%d)\n",
+                    rank, nNicsToMerge, NCCL_NET_MAX_DEVS_PER_NIC);
+            return -1;
+        }
+
+        int outMergedDev = -1;
+
+        int ndev = 0;
+        RCCL_TEST_CHECK(GetDeviceCount(&ndev));
+        if (ndev <= 0) return -1;
+
+        std::vector<ncclNetProperties_t> props(ndev);
+        std::vector<int> physDevs;
+        for (int i = 0; i < ndev; i++) {
+            memset(&props[i], 0, sizeof(ncclNetProperties_t));
+            RCCL_TEST_CHECK(GetDeviceProperties(i, &props[i]));
+            if (!props[i].name || !strchr(props[i].name, '+'))
+                physDevs.push_back(i);
+        }
+
+        if (physDevs.size() < offset + 1) {
+            return -1;
+        }
+
+        int targetSpeed = props[physDevs[offset]].speed;
+        std::vector<int> compat;
+        for (int d : physDevs)
+            if (props[d].speed == targetSpeed) compat.push_back(d);
+
+        if ((int)compat.size() < nNicsToMerge * 2) {
+            return CreateMergedDevice(nNicsToMerge, rank, offset + 1);
+        }
+
+        int baseOffset = (rank % 2) ? nNicsToMerge : 0;
+        int finalOffset = baseOffset + offset;
+
+        if (finalOffset + nNicsToMerge > (int)compat.size()) return -1;
+
+        ncclNetVDeviceProps_t vProps;
+        memset(&vProps, 0, sizeof(vProps));
+        vProps.ndevs = nNicsToMerge;
+        for (int i = 0; i < nNicsToMerge; i++)
+            vProps.devs[i] = compat[finalOffset + i];
+
+        if (MakeVirtualDevice(&outMergedDev, &vProps) != ncclSuccess || outMergedDev < 0) {
+            fprintf(stderr,
+                    "Rank %d failed to create %d-NIC merged device at offset %d, retrying with offset %d\n",
+                    rank, nNicsToMerge, offset, offset + 1);
+            return CreateMergedDevice(nNicsToMerge, rank, offset + 1);
+        }
+
+        return outMergedDev;
+    }
+};
+
+#endif // MPI_TESTS_ENABLED
